+++ src/app/search/search-filter.component.js
@@ -15,10 +15,12 @@ export const SearchFilterComponent = {
 };
 
 function SearchFilterCtrl(
-  $scope, dimStoreService, D2StoresService, dimVendorService, dimSearchService, hotkeys, $i18next, $element, dimCategory, D2Categories, dimSettingsService, toaster, ngDialog) {
+  $scope, dimStoreService, D2StoresService, dimVendorService, dimSearchService, dimItemInfoService, hotkeys, $i18next, $element, dimCategory, D2Categories, dimSettingsService, toaster, ngDialog) {
   'ngInject';
   const vm = this;
   vm.search = dimSearchService;
+  vm.bulkItemTags = _.clone(dimSettingsService.itemTags);
+  vm.bulkItemTags.push({ type: 'clear', label: 'Tags.ClearTag' });
 
   function getStoreService() {
     return vm.destinyVersion === 2 ? D2StoresService : dimStoreService;
@@ -26,6 +28,7 @@ function SearchFilterCtrl(
 
   let filters;
   let searchConfig;
+  let filteredItems = [];
 
   vm.$onChanges = function(changes) {
     if (changes.destinyVersion && changes.destinyVersion) {
@@ -163,12 +166,27 @@ function SearchFilterCtrl(
   };
 
   vm.clearFilter = function() {
+    filteredItems = [];
     vm.search.query = "";
     vm.filter();
     textcomplete.trigger('');
   };
 
+  vm.bulkTag = function() {
+    dimItemInfoService.bulkSave(filteredItems.filter((i) => i.taggable).map((item) => {
+      item.dimInfo.tag = vm.selectedTag.type === 'clear' ? undefined : vm.selectedTag.type;
+      return item;
+    }));
+
+    // invalidate and filter
+    filters.reset();
+    vm.filter();
+    vm.showSelect = false;
+  };
+
   vm.filter = function() {
+    vm.selectedTag = undefined;
+    filteredItems = [];
     let filterValue = (vm.search.query) ? vm.search.query.toLowerCase() : '';
     filterValue = filterValue.replace(/\s+and\s+/, ' ');
 
@@ -176,6 +194,9 @@ function SearchFilterCtrl(
 
     for (const item of getStoreService().getAllItems()) {
       item.visible = filterFn(item);
+      if (item.visible) {
+        filteredItems.push(item);
+      }
     }
 
     if (vm.destinyVersion === 1) {